[CSE] Major code cleanup and refactoring; the editorAPI acts as the primary interface to the CS' code Added BaseFormComponent CSAS script commands Updated muParserX lib to 1.11, which fixes all outstanding bugs Added CSAS documentation generation code CSAS scripts that accept parameters will now require a valid parameter list passed to them during execution Made CSAS global variable type distinguishable in the management dialog Fixed a bug that triggered a deadlock when the legacy CS crashed during lip sync generation Added a new console command: DumpDocs Exposed/updated the following editor class definitions: TESModelAnim, TES, FileFinder, BSTextureManager, DialogExtraSubWindow, DialogExtraWorkingData, Script, ExtraDataList, TESClimate, TESEffectShader, TESForm, TESLoadScreen, TESObjectListHead, TESObjectANIO, TESObjectCELL, TESObjectREFR, TESQuest, TESSound, TESWorldSpace, NiObjectNET, Archive, ArchiveFile, CompressedArchiveFile, BSArchive, BSArchiveHeader, BSFile, NiBinaryStream, NiFile, TESDialog Result script compilation ops generate a notification when failed Improved object window refresh performance Added new filter tools to the Object and Cell View windows Added the ability to change the active landscape texture from the 'Landscape Texture Use' dialog Added the ability to edit result scripts in a larger text edit field Fixed a CS bug that progressively reduced the size of the Cell View window's controls after each plugin load Added the ability to sample sound files inside BSA archives without having to extract them Fixed a bug in the actor preview control hook that prevented animations from being previewed Added various "achievements" to the CS Added diagnostic messages to TESDialog::BuildSubWindow The INIManager class now releases allocated INISettings during deinit Added top-level crash handling for the CS; unhandled exceptions will create a minidump before crashing Fixed a memory leak in the HallOfFame code Fixed a CS bug that would cause a CTD when quick-loading plugins with TESObjectREFR records with dangling base forms Added a toggle to prevent plugin save operations from modifying their file timestamps Removed a couple of redundant items for the main menu Added a new sub-menu to the 'File' menu Rewrote the 'Copy Path' dialog's mouse capture code Added a 'Jump to Cell Coords' tool to the Cell View window, a la the GECK Fixed a bug that occasionally caused a CTD when using the 'Copy External File' tool in the response window Added error handling for the lip sync generator code Default CS/CSE directories are now created on CS startup Fixed a bug in the asset selector code that incorrectly handled idle anim file paths Added code to all CLR entry points to set the default thread apartment state to STA Many minor changes and fixes
[SE]
Made save confirmation dialog boxes use the 'Topmost' window style
Fixed IntelliSense positioning during editor window first-run, move/resize ops
Added new icons for the script type and Edit dropdown lists
Fixed a bug that prevented IntelliSense tooltips from being displayed when the text edit field wasn't in focus
Fixed various positioning issues with IntelliSense tooltips
Minor changes and fixes

[BSA]
Minor changes to the UI

[TAG]
Added support for all form types and drop locations

[*]
Build fixes
Updated TODO
Updated readMe

+++ [Components]/ScriptEditor/ScriptEditorManager.cpp
@@ -27,7 +27,7 @@ ScriptEditorManager^% ScriptEditorManager::GetSingleton()
 
 void ScriptEditorManager::PerformOperation(ScriptEditorManager::OperationType Op, ScriptEditorManager::OperationParams^ Parameters)
 {
-	try 
+	try
 	{
 		UInt16 NoOfParams = Parameters->ParameterList->Count, ParamsRequired = ParamCount[(int)Op];
 
@@ -159,9 +159,9 @@ ScriptEditor::Workspace^ ScriptEditorManager::GetAllocatedWorkspace(UInt32 Alloc
 {
 	ScriptEditor::Workspace^ Result = nullptr;
 
-	for each (ScriptEditor::Workspace^% Itr in WorkspaceAllocationMap) 
+	for each (ScriptEditor::Workspace^% Itr in WorkspaceAllocationMap)
 	{
-		if (Itr->GetAllocatedIndex() == AllocatedIndex) 
+		if (Itr->GetAllocatedIndex() == AllocatedIndex)
 		{
 			Result = Itr;
 			break;
@@ -176,7 +176,7 @@ bool ScriptEditorManager::MoveScriptDataToVanillaEditor(ScriptEditor::Workspace^
 	{
 		CStringWrapper^ CScriptText = gcnew CStringWrapper(PreprocessedScriptResult);
 		g_ScriptDataInteropPackage->Text = CScriptText->String();
-		g_ScriptDataInteropPackage->Type = CSEEditor->GetScriptType();
+		g_ScriptDataInteropPackage->Type = (UInt16)CSEEditor->GetScriptType();
 		g_ScriptDataInteropPackage->ModifiedFlag = CSEEditor->GetModifiedStatus();
 		NativeWrapper::ScriptEditor_SetScriptData(CSEEditor->GetAllocatedIndex(), g_ScriptDataInteropPackage);
 		return true;
